correct strncmp prototype

+++ ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: afelger <afelger@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 14:14:18 by afelger           #+#    #+#             */
-/*   Updated: 2024/10/17 14:05:02 by afelger          ###   ########.fr       */
+/*   Updated: 2024/10/18 16:40:14 by afelger          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long	ft_strncmp(const char *s1, const char *s2, unsigned long n)
+int	ft_strncmp(const char *s1, const char *s2, unsigned long n)
 {
 	unsigned long		counter;
 	const unsigned char	*str1;
